Use int16_t for charge

+++ NtupleTools/src/NtupleGenParticles.cc
@@ -10,42 +10,42 @@ NtupleGenParticles::NtupleGenParticles(const edm::ParameterSet& iConfig) :
 
     token_ = consumes<reco::GenParticleCollection>(inputTag_);
 
-    produces<std::vector<float> > (prefix_ + "px"    + suffix_);
-    produces<std::vector<float> > (prefix_ + "py"    + suffix_);
-    produces<std::vector<float> > (prefix_ + "pz"    + suffix_);
-    produces<std::vector<float> > (prefix_ + "E"     + suffix_);
-    produces<std::vector<float> > (prefix_ + "pt"    + suffix_);
-    produces<std::vector<float> > (prefix_ + "eta"   + suffix_);
-    produces<std::vector<float> > (prefix_ + "phi"   + suffix_);
-    produces<std::vector<float> > (prefix_ + "M"     + suffix_);
-    produces<std::vector<float> > (prefix_ + "vx"    + suffix_);
-    produces<std::vector<float> > (prefix_ + "vy"    + suffix_);
-    produces<std::vector<float> > (prefix_ + "vz"    + suffix_);
-    produces<std::vector<int> >   (prefix_ + "charge"+ suffix_);
-    produces<std::vector<int> >   (prefix_ + "pdgId" + suffix_);
-    produces<std::vector<int> >   (prefix_ + "status"+ suffix_);
-    produces<unsigned>            (prefix_ + "size"  + suffix_);
+    produces<std::vector<float> >   (prefix_ + "px"    + suffix_);
+    produces<std::vector<float> >   (prefix_ + "py"    + suffix_);
+    produces<std::vector<float> >   (prefix_ + "pz"    + suffix_);
+    produces<std::vector<float> >   (prefix_ + "E"     + suffix_);
+    produces<std::vector<float> >   (prefix_ + "pt"    + suffix_);
+    produces<std::vector<float> >   (prefix_ + "eta"   + suffix_);
+    produces<std::vector<float> >   (prefix_ + "phi"   + suffix_);
+    produces<std::vector<float> >   (prefix_ + "M"     + suffix_);
+    produces<std::vector<float> >   (prefix_ + "vx"    + suffix_);
+    produces<std::vector<float> >   (prefix_ + "vy"    + suffix_);
+    produces<std::vector<float> >   (prefix_ + "vz"    + suffix_);
+    produces<std::vector<int16_t> > (prefix_ + "charge"+ suffix_);
+    produces<std::vector<int> >     (prefix_ + "pdgId" + suffix_);
+    produces<std::vector<int> >     (prefix_ + "status"+ suffix_);
+    produces<unsigned>              (prefix_ + "size"  + suffix_);
 }
 
 NtupleGenParticles::~NtupleGenParticles() {}
 
 void NtupleGenParticles::produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
 
-    std::auto_ptr<std::vector<float> > v_px    (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_py    (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_pz    (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_E     (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_pt    (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_eta   (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_phi   (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_M     (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_vx    (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_vy    (new std::vector<float>());
-    std::auto_ptr<std::vector<float> > v_vz    (new std::vector<float>());
-    std::auto_ptr<std::vector<int> >   v_charge(new std::vector<int>());
-    std::auto_ptr<std::vector<int> >   v_pdgId (new std::vector<int>());
-    std::auto_ptr<std::vector<int> >   v_status(new std::vector<int>());
-    std::auto_ptr<unsigned>            v_size  (new unsigned(0));
+    std::auto_ptr<std::vector<float> >   v_px    (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_py    (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_pz    (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_E     (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_pt    (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_eta   (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_phi   (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_M     (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_vx    (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_vy    (new std::vector<float>());
+    std::auto_ptr<std::vector<float> >   v_vz    (new std::vector<float>());
+    std::auto_ptr<std::vector<int16_t> > v_charge(new std::vector<int16_t>());
+    std::auto_ptr<std::vector<int> >     v_pdgId (new std::vector<int>());
+    std::auto_ptr<std::vector<int> >     v_status(new std::vector<int>());
+    std::auto_ptr<unsigned>              v_size  (new unsigned(0));
 
     //__________________________________________________________________________
     if (!iEvent.isRealData()) {
